move parse_env to ddcommon
