+++ ddcommon/Cargo.toml
@@ -13,20 +13,30 @@ crate-type = ["lib"]
 [dependencies]
 anyhow = "1.0"
 futures = "0.3"
-futures-core = {version = "0.3.0", default-features = false}
-futures-util = {version = "0.3.0", default-features = false}
+futures-core = { version = "0.3.0", default-features = false }
+futures-util = { version = "0.3.0", default-features = false }
 hex = "0.4"
-hyper = {version = "0.14", features = ["http1", "client", "tcp", "stream"], default-features = false}
-hyper-rustls = {version = "0.23", default-features = false, features = ["native-tokio", "http1", "tls12"]}
+http = "0.2"
+hyper = { version = "0.14", features = [
+    "http1",
+    "client",
+    "tcp",
+    "stream",
+], default-features = false }
+hyper-rustls = { version = "0.23", default-features = false, features = [
+    "native-tokio",
+    "http1",
+    "tls12",
+] }
 lazy_static = "1.4"
-log = {version = "0.4"}
+log = { version = "0.4" }
 pin-project = "1"
 regex = "1.5"
-rustls = {version = "0.20.4", default-features = false}
-rustls-native-certs = {version = "0.6"}
-tokio = {version = "1.23", features = ["rt", "macros"]}
-tokio-rustls = {version = "0.23"}
-serde = {version = "1.0", features = ["derive"]}
+rustls = { version = "0.20.4", default-features = false }
+rustls-native-certs = { version = "0.6" }
+tokio = { version = "1.23", features = ["rt", "macros"] }
+tokio-rustls = { version = "0.23" }
+serde = { version = "1.0", features = ["derive"] }
 
 [dev-dependencies]
 indexmap = "1.8"
